requirements.txt
openai-whisper
pydub
pyaudio
keyboard
python-dotenv
openai
requests